[PR] addressed comments
